procfs/pool_state: Set TESTPOOL2 after scsi setup
The scsi_debug kernel module is used to simulate device failures on
Linux.  Testbeds may not have this module, and should gracefully SKIP
this test in that contingency.  This is performed by the load_scsi_debug
macro.  However, after a SKIP is registered there, the local cleanup
macro is triggered.  On this test, this macro assumes that TESTPOOL2
being nonempty means that a pool of that name has been created, and
should be destroyed.

Currently, TESTPOOL2 is set _before_ load_scsi_debug, so failures in
that call will trigger an attempt to clean up a nonexistent pool.  The
SKIP is promoted to a FAIL, and leading to erroneous ZTS failures on,
e.g., Ubuntu 20.04.

This patch sets TESTPOOL2 _after_ load_scsi_debug.

Signed-off-by: Antonio Russo <aerusso@aerusso.net>